fix!: remove cache components config for gulp runtime
When you start gulp scripts, in node process created components
config cache. But, if create the component with an artisan in another
terminal, the cache from the first node process will be not updated
and the component will be added to the 'none' namespace.

BREAKING CHANGE: in config.ts removed checkConfigIntegrity and not added
to the components config component will be not added automatic.